==> 2b. GWR & MGWR.ipynb <==
!pip install geopandas
!pip install mgwr
!pip install numba

import numpy as np
import pandas as pd
import pickle
import os
import multiprocessing
os.environ['USE_PYGEOS'] = '0'
import geopandas as gp
import multiprocessing as mp
from mgwr.gwr import GWR,MGWR
from mgwr.sel_bw import Sel_BW

#Load the funda data
funda_data = gp.read_file("data/Houseprices/funda_buy_amsterdam_31-03-2023_full_distances.gpkg")

# funda_data = funda_data.loc[funda_data['price'] < 2500000]
# funda_data = pd.DataFrame(funda_data)

# funda_data = gp.GeoDataFrame(pd.get_dummies(funda_data, columns = ["house_type", "building_type", "energy_label", "has_balcony", "has_garden"]))

#show funda data
funda_data


#create array with the dependent variable
b_y = funda_data['price_m2'].values.reshape((-1,1))

# #create an array with the indepentend variables (order matters for the extraction of params later)
# b_X = funda_data[['room', 'bedroom', 'bathroom', 'living_area',
#                   'house_age', 'bus_dist', 'subway_dist', 'train_dist',
#                   'university_dist', 'school_dist', 'mall_dist', 'supermarket_dist',
#                   'energy_label_A', 'energy_label_B', 'energy_label_C', 'energy_label_D',
#                   'energy_label_E', 'energy_label_F', 'energy_label_G', 'energy_label_na',
#                   'has_balcony_0.0', 'has_balcony_1.0', 'has_garden_0.0', 'has_garden_1.0']].values 

# cols = ['room', 'bedroom', 'bathroom', 'living_area',
#                   'house_age', 'train_dist', 'metro_dist', 'tram_dist', 'ndvi100']

cols = ['living_area', 'bedroom', 'bathroom',
                  'house_age', 'train_dist', 'metro_dist', 'tram_dist', 'ndvi300', 'centre_dist', 'zuid_dist', 'shops_dist', 'school_dist']

b_X = funda_data[cols].values 

#create coordinate tuple for the model
u = funda_data['geometry'].x
v = funda_data['geometry'].y
b_coords = list(zip(u, v))

multiprocessing.cpu_count()

#This might be needed to turn off the OpenMP multi-threading
%env OMP_NUM_THREADS = 1

#Parrallelization is more favored when you your data are large and/or your machine have many many cores.
#mgwr has soft dependency of numba, please install numba if you need better performance (pip install numba).
n_proc = 4 #two processors
pool = mp.Pool(n_proc) 

%%time
#Run basic GWR in parrallel mode

gwr_selector = Sel_BW(b_coords, b_y, b_X, kernel='bisquare')
gwr_bw = gwr_selector.search(pool = pool, criterion = "AIC") #add pool to Sel_BW.search
print(gwr_bw)
gwr_results = GWR(b_coords, b_y, b_X, gwr_bw, kernel="bisquare").fit(pool = pool)

gwr_results.summary()

%%time
#run MGWR in parrallel mode. Note: max_iter_multi needs to be specified

mgwr_selector = Sel_BW(b_coords, b_y, b_X, multi=True, kernel = 'bisquare')
mgwr_bw = mgwr_selector.search(pool=pool, max_iter_multi=50, criterion = "AIC", verbose = True, init_multi = gwr_bw) #add pool to Sel_BW.search
print(mgwr_bw)


#pickle.dump(mgwr_selector, open('data/models/mgwr_selector_ndvi_m2.pkl', 'wb'))

# mgwr_selector = pickle.load(open('data/models/mgwr_selector_ndvi_m2.pkl', 'rb'))

mgwr_results = MGWR(b_coords, b_y, b_X, selector=mgwr_selector).fit(pool = pool)


#show MGWR model summary
mgwr_results.summary()

#recreate R format table

def return_geopackage(mgwr_results,cols, path_name):
    df = gp.GeoDataFrame()
    df['Intercept'] = mgwr_results.params[:,0]
    df['intercept_SE'] = mgwr_results.bse[:,0]
    df['intercept_TV'] = mgwr_results.tvalues[:,0]
    
    df['yhat'] = mgwr_results.predy
    df['residual'] = mgwr_results.y.reshape((-1,1)) - mgwr_results.predy
    
    for i,col in enumerate(cols):
        
        df[col] = mgwr_results.params[:,i+1]

        df[col+'_SE'] = mgwr_results.bse[:,i+1]
        df[col+'_TV'] = mgwr_results.tvalues[:,i+1]
        
    df['geometry'] = funda_data['geometry']
    df.to_file(path_name)

#show the GeoDataFrame and write it to a folder
return_geopackage(gwr_results,cols, 'data/models/gwr_results_amsterdam_ndvi300_m2.gpkg')

#pickle.dump(mgwr_results, open('data/models/mgwr_results_amsterdam_ndvi_m2.pkl', 'wb'))

#gwr_results = pickle.load(open('data/models/mgwr_results_amsterdam_ndvi_m2.pkl', 'rb'))
